Reorganise entrypoints into specific tasks
