+++ apps/api/package.json
@@ -9,7 +9,10 @@
     "prebuild": "rimraf dist",
     "build": "nest build",
     "build:types": "tsc -b tsconfig.types.json ",
-    "repl": "nest start --entryFile repl",
+    "start:repl": "nest start --entryFile entrypoints/repl/main",
+    "start:api": "nest start --entryFile entrypoints/api/main",
+    "start:react-static": "nest start --entryFile entrypoints/reactStatic/main",
+    "start:react-ssr": "nest start --entryFile entrypoints/reactSSR/main",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "dev": "nest build --webpack --webpackPath webpack-hmr.config.js --watch",
