--- apps/api/src/controllers/app.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { AppService } from '../app.service';
-
-@Controller('/api')
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  async getHello(): Promise<{ message: string }> {
-    return await this.appService.getHello();
-  }
-}
