--- apps/api/src/repl.ts
@@ -1,7 +1,0 @@
-import { AppModule } from './app.module';
-import { repl } from '@nestjs/core';
-
-async function bootstrap() {
-  await repl(AppModule);
-}
-bootstrap();
