+++ apps/api/src/services/hello.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AppService {
+export class HelloService {
   async getHello(): Promise<{ message: string }> {
     return { message: 'Hello World' };
   }
