+++ apps/api/src/services/services.module.ts
@@ -1,14 +1,16 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './userService';
+import { UserService } from './user.service';
 import { RepositoriesModule } from 'src/repository/repositories.module';
+import { HelloService } from './hello.service';
 
 @Module({
   imports: [RepositoriesModule],
-  providers: [UserService],
-  exports: [UserService],
+  providers: [UserService, HelloService],
+  exports: [UserService, HelloService],
 })
-export class ServicesModule {}
+export class ServicesModule { }
 
 export type Services = {
   userService: UserService;
+  helloService: HelloService;
 };
